Added (float) cast to ceil call to avoid problem in VC.NET.
git-svn-id: https://svn.code.sf.net/p/freemat/code/trunk@347 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ FreeMat2/libs/libCore/IO.cpp
@@ -482,7 +482,7 @@ namespace FreeMat {
       fseek(fptr->fp,0L,SEEK_END);
       fsize = ftell(fptr->fp) - fcpos + 1;
       fseek(fptr->fp,fcpos,SEEK_SET);
-      dp[infiniteDim] = ceil(fsize/elementSize/elementCount);
+      dp[infiniteDim] = ceil((float)(fsize/elementSize/elementCount));
       elementCount *= (int) dp[infiniteDim];
     }
     // Next, we allocate space for the result
